Refactor Superduper mode
