+++ users/narze/superduper.h
@@ -1,0 +1,7 @@
+#pragma once
+#include "narze.h"
+
+void toggle_superduper_mode(void);
+void set_superduper_key_combo_layer(uint16_t layer);
+void set_superduper_key_combos(void);
+void clear_superduper_key_combos(void);
